refactor: split routes and add utils
